Split capcom package in files related to concerns
